Prevent read methods from succeeding on partial reads
Fixes the following tests for Go:

- [EofExceptionU4](https://github.com/kaitai-io/kaitai_struct_tests/blob/3c5fb218ba78fb040243b5a35cea69fb974118cd/formats/eof_exception_u4.ksy)
- [EosExceptionU4](https://github.com/kaitai-io/kaitai_struct_tests/blob/3c5fb218ba78fb040243b5a35cea69fb974118cd/formats/eos_exception_u4.ksy)

The failures of the above tests in CI have been indicating for a while
that the read primitives in the Go runtime library succeed even if there
are not enough bytes available in the stream. More specifically, they
only fail if we are right at the end of file and there are 0 bytes
available, but they apparently never return an error if there is at
least 1 byte remaining regardless of how many bytes the particular read
primitive actually needs.

The problem is using the `io.Reader.Read` function, see
https://pkg.go.dev/io#Reader (**emphasis** mine):

> Read reads **up to** len(p) bytes into p. (...) If some data is
> available but not len(p) bytes, Read conventionally **returns what is
> available instead of waiting for more**.

> When Read encounters an error or end-of-file condition after
> successfully reading n > 0 bytes, it returns the number of bytes read.
> It may return the (non-nil) error from the same call or return the
> error (and n == 0) from a subsequent call. An instance of this general
> case is that a Reader returning a non-zero number of bytes at the end
> of the input stream may return either err == EOF **or err == nil**.
> The next Read should return 0, EOF.

So this function is specifically allowed to read fewer bytes than
requested and this situation is not guaranteed to be indicated by an
error. This is exactly what we saw in the tests.

That's why we're now using
[`io.ReadFull`](https://pkg.go.dev/io#ReadFull) instead:

> ReadFull reads **exactly len(buf) bytes** from r into buf. It returns the
> number of bytes copied and an error if fewer bytes were read. The
> error is EOF only if no bytes were read. If an EOF happens after
> reading some but not all the bytes, ReadFull returns ErrUnexpectedEOF.
> On return, **n == len(buf) if and only if err == nil**.

The distinction between EOF and ErrUnexpectedEOF is a bit weird, we
generally have only one kind of EOF error in other languages, but let's
not worry about that now. At least it fixes the primary bug.